Phase 1 prep: Float->Numeric money-column migration (AUDIT risk #4)
Storage-only, behaviour-preserving retype of the 28 Float money/money-adjacent
columns to fixed-scale Numeric. NOT yet run against any database, NOT merged.

Models (Float -> Numeric), scales confirmed with the founder:
- prices (line unit_price, catalogue purchase/sale)  -> Numeric(10,4)
- quantities (line qty, stock levels)                -> Numeric(10,3)
- line amounts + supplier-invoice tax amounts        -> Numeric(10,2)
- document totals / payment amounts                  -> Numeric(12,2)
- rates / markup percentages                         -> Numeric(5,2)
Across: invoice, customer_invoice(+line), customer_quote(+line),
customer_payment, customer_invoice_payment, product_service, user,
product_cache. Geometry Floats in takeoff (pixels/sqm) deliberately left as
Float - not money.

Migration migrations/versions/money_float_to_numeric.py:
- Hand-written (not autogenerated, to avoid sweeping up the risk #10 runtime
  schema drift). down_revision = add_job_cards; single head confirmed via
  flask db heads.
- batch_alter_table for SQLite-safety; postgresql_using 'col::numeric' so the
  target scale rounds existing values deterministically. Full downgrade() back
  to Float included.

NOTE - planned Phase 0 placeholder deletion was abandoned: investigation flagged
migration_add_imap_fields.py as a duplicate, but flask db heads proved it is the
real 'add_imap_fields' migration (merge_heads_001 depends on it). Deleting it
broke the graph (KeyError: 'add_imap_fields'); it was restored and is untouched
here. This commit is purely about money columns.

This is Phase 1 (storage). Phase 2 (Decimal money() helper + converting the
float round() calc/sync sites + test corpus) is a separate later change; risk #4
is not closed by this commit alone. Execution next session: prod/staging
flask db current check -> staging rehearsal on a prod-data copy (diff before/after,
flag any row delta > GBP 0.005) -> production deploy with a fresh backup.

+++ app/models/customer_payment.py
@@ -9,7 +9,7 @@ class CustomerPayment(db.Model):
     id = db.Column(db.Integer, primary_key=True)
     user_id = db.Column(db.Integer, db.ForeignKey('user.id'), nullable=False, index=True)
     customer_id = db.Column(db.Integer, db.ForeignKey('customer.id'), nullable=False)
-    amount = db.Column(db.Float, nullable=False)
+    amount = db.Column(db.Numeric(12, 2), nullable=False)
     payment_date = db.Column(db.Date, nullable=False, default=datetime.utcnow)
     payment_method = db.Column(db.String(50), default='bank_transfer')
     reference = db.Column(db.String(255), nullable=True)
@@ -42,6 +42,6 @@ class CustomerInvoicePayment(db.Model):
     id = db.Column(db.Integer, primary_key=True)
     payment_id = db.Column(db.Integer, db.ForeignKey('customer_payment.id'), nullable=False)
     invoice_id = db.Column(db.Integer, db.ForeignKey('customer_invoice.id'), nullable=False)
-    amount_applied = db.Column(db.Float, nullable=False)
+    amount_applied = db.Column(db.Numeric(12, 2), nullable=False)
 
     invoice = db.relationship('CustomerInvoice', backref='payment_links', lazy='select')
